requirements.txt
matplotlib==3.8.0
pandas==2.1.1
aio-binance-library==2.0.7
scipy==1.11.3

